Fixed Performance loss after fastai update -> Changed Callback method names

+++ seqdata/core.py
@@ -207,7 +207,7 @@ class HDF2Sequence(Transform):
             seq = seq[truncate_sz:]
 
         #it is important to slice first and then do the class conversion
-        return self.to_cls(seq.astype('f8'))#workaround for random bug, that mitigates convergence if the numpy array is an f4 array. Seems to make no sense because the result does not change.
+        return self.to_cls(seq)#workaround for random bug, that mitigates convergence if the numpy array is an f4 array. Seems to make no sense because the result does not change.
 
     def encodes(self, item)->None:
         return self._extract_dict_sequence(item)
